+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for Cloud Fetch (#146, #151, #154)
 - SQLAlchemy has_table function now honours schema= argument and adds catalog= argument (#174)
+- SQLAlchemy set non_native_boolean_check_constraint False as it's not supported by Databricks (#120)
 - Fix: Revised SQLAlchemy dialect and examples for compatibility with SQLAlchemy==1.3.x (#173)
 - Fix: oauth would fail if expired credentials appeared in ~/.netrc (#122)
 - Fix: Python HTTP proxies were broken after switch to urllib3 (#158)